Main valentin

+++ backend/src/main/java/com/hs_esslingen/insy/dto/PriceDTO.java
@@ -1,7 +1,5 @@
 package com.hs_esslingen.insy.dto;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PriceDTO {
-    private BigDecimal maxPrice;
-    private BigDecimal minPrice;
+    private Integer maxPrice;
+    private Integer minPrice;
 }
